requirements.txt
joblib
scipy
tables
optuna
mlflow
keras
dm-sonnet
seaborn
matplotlib
tqdm
tensorflow-gpu
numpy
scikit-learn
pandas
click